Add el_base32_{decode,encode} C functions.
Add base32 implementation from
http://code.google.com/p/google-authenticator/source/browse/libpam/ to
the C library.

License is the Apache License as well.  It is short, easy to understand
C code and unlike EL's current implementation, it handles extra
whitespace or dashes and common typos during decoding.

No changes to it, except for adding el_ prefix to function names and
improving docstrings.