Use truncated SHA-256 instead of SHA-1.
Using truncated output of one of the SHA-2 function is recommended over
the vulnerable SHA-1 hash and is supposed to be stronger.

See
http://crypto.stackexchange.com/questions/9435/is-truncating-a-sha512-hash-to-the-first-40-characters-as-secure-as-using-sha1
for a more detailed discussion.

Notice that blocked keys are still hashed with (full) SHA-1: it is
shorter than SHA-256, we don't worry about collisions in this use case
and it doesn't hurt if the digest hash and blocklist hash are different.